requirements.txt
coverage==3.5.3
django-discover-runner==0.2.2
flake8==1.5
mock==1.0.0
suds==0.4
